requirements.txt
cupy-cuda12x>=13.3.0
matplotlib>=3.9.2
